+++ CheckVersion/README.md
@@ -32,13 +32,13 @@ Notice we could have adopted an opt-in scheme to make "just include those files"
 
 ### `.cv`
 
-cv saves history inside the `.cv` folder. File update time is stored as utc.
+CheckVersion saves history inside the `.cv` folder. File update time is stored as utc.
 
 ### `.cvignore`
 
-cv uses a `.cvignore` file, which shares the same `.gitignore` file format as `git`, this is used to decide which files cv should consider when issuing `status` command; Notice we are using a new file name instead of using .gitignore directly - this is to allow the same place having a git repo.
+CheckVersion uses a `.cvignore` file, which shares the same `.gitignore` file format as `git`, this is used to decide which files CheckVersion should consider when issuing `status` command; Notice we are using a new file name instead of using .gitignore directly - this is to allow the same place having a git repo.
 
-At the moment cv just checks whether the beginning of paths match that as specified in .cvignore file - no wildcards is supported. You are welcome to make a PR in [this function](https://github.com/chaojian-zhang/cv/blob/91f711abcf1ba6d6a37ab8d3dc9c2d79ee694cc9/Program.cs#L344) to complete the implementation.
+At the moment CheckVersion just checks whether the beginning of paths match that as specified in .cvignore file - no wildcards is supported. You are welcome to make a PR in [this function](https://github.com/chaojian-zhang/CheckVersion/blob/91f711abcf1ba6d6a37ab8d3dc9c2d79ee694cc9/Program.cs#L344) to complete the implementation.
 
 ## Examples
